requirements.txt
# SPDX-FileCopyrightText: Copyright (c) 2025 Broadsage <opensource@broadsage.com>
#
# SPDX-License-Identifier: Apache-2.0

# Python dependencies for docker-scaffold

# Required for YAML configuration parsing
PyYAML>=6.0.1

# Required for format-preserving YAML updates
ruamel.yaml>=0.18.0

# Required for executing Ansible playbooks
ansible>=2.20.0
